requirements.txt
Django==1.6
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
psycopg2==2.5.1
django-tastypie==0.10.0
python-mimeparse==0.1.4
